+++ libs/base_core/src/data/error/mod.rs
@@ -1,0 +1,16 @@
+// devela_base_core::data::error
+//
+//! Data-related errors.
+//
+
+mod capacity;
+mod other;
+
+crate::structural_mods! { // _mods
+    _mods {
+        pub use super::{
+            capacity::*,
+            other::*,
+        };
+    }
+}